Add method to disable features within modules
