+++ backend/dataall/core/permission_checker.py
@@ -6,6 +6,7 @@ from typing import Protocol, Callable
 
 from dataall.core.context import RequestContext, get_context
 from dataall.db.api import TenantPolicy, ResourcePolicy, Environment
+from dataall.utils.decorator_util import process_func
 
 
 class Identifiable(Protocol):
@@ -48,21 +49,6 @@ def _check_resource_permission(session, uri, permission):
     )
 
 
-def _process_func(func):
-    """Helper function that helps decorate methods/functions"""
-    def no_decorated(f):
-        return f
-
-    static_func = False
-    try:
-        fn = func.__func__
-        static_func = True
-    except AttributeError:
-        fn = func
-
-    # returns a function to call and static decorator if applied
-    return fn, staticmethod if static_func else no_decorated
-
 
 def has_resource_permission(
         permission: str,
@@ -80,7 +66,7 @@ def has_resource_permission(
         param_name = "uri"
 
     def decorator(f):
-        fn, fn_decorator = _process_func(f)
+        fn, fn_decorator = process_func(f)
 
         def decorated(*args, **kwargs):
             uri: str
@@ -114,7 +100,7 @@ def has_tenant_permission(permission: str):
     All the information about the user is retrieved from RequestContext
     """
     def decorator(f):
-        fn, fn_decorator = _process_func(f)
+        fn, fn_decorator = process_func(f)
 
         def decorated(*args, **kwargs):
             with get_context().db_engine.scoped_session() as session:
@@ -129,7 +115,7 @@ def has_tenant_permission(permission: str):
 
 def has_group_permission(permission):
     def decorator(f):
-        fn, fn_decorator = _process_func(f)
+        fn, fn_decorator = process_func(f)
 
         def decorated(*args, admin_group, uri, **kwargs):
             with get_context().db_engine.scoped_session() as session:
